What this repository has been made for
The repository consists of all the projects and programs I have made during my studying for the sake of improving my coding skills and standard. The projects and programs are not written in a specific language. Some of them are Java, and some are C#. In the future, some other languages may be used.